Added compiler directives to not negatively impact the runtime of meshes with uniform subdivision levels.
Ray tracing meshes with triangles that have varying subdivision levels is a bit slower, because it needs to compute more stuff and suffers more conditional branching. All this is not necessary when ray tracing meshes with a uniform subdivision level.

So we use compiler directives to "disable" this expensive code when meshes are run with a uniform subdivision level.